MVP/ai_stack/pyproject.toml
[project]
name = "world-of-shadows-ai-stack"
version = "1.0.0"
description = "AI stack: capabilities registry, MCP surface, and runtime integration for World of Shadows"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
]

[project.optional-dependencies]
retrieval = [
    "numpy>=1.24.0,<3",
    "fastembed>=0.3.0,<1",
    "langchain>=1.2.0,<2",
    "langchain-core>=1.2.0,<2",
]
langchain = [
    "langchain>=1.2.0,<2",
    "langchain-core>=1.2.0,<2",
    "langgraph>=1.0.3,<2",
]
# Everything needed for `pytest ai_stack/tests` (LangGraph + RAG + YAML GoC loaders).
# Without these, `langgraph_runtime` import fails and `ai_stack/__init__.py` omits
# RuntimeTurnGraphExecutor from `__all__` by design (MCP / lightweight import path).
test = [
    "pytest>=7.0,<9",
    "pytest-asyncio>=0.21,<1",
    "pytest-cov>=4.0,<6",
    "pytest-timeout>=2.1",
    "langchain>=1.2.0,<2",
    "langchain-core>=1.2.0,<2",
    "langgraph>=1.0.3,<2",
    "numpy>=1.24.0,<3",
    "fastembed>=0.3.0,<1",
    "pyyaml>=6.0,<7",
    "httpx>=0.27.0,<1",
]
all = [
    "numpy>=1.24.0,<3",
    "fastembed>=0.3.0,<1",
    "langchain>=1.2.0,<2",
    "langchain-core>=1.2.0,<2",
    "langgraph>=1.0.3,<2",
    "pytest>=7.0,<9",
    "pytest-asyncio>=0.21,<1",
    "pytest-cov>=4.0,<6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
